refactor(recruit): organize rich text editor toolbar into own component
