+++ sites/recruit.sveltevietnam.dev/src/lib/components/rich-text-editor/RichTextEditor.svelte
@@ -1,19 +1,12 @@
 <script lang="ts" module>
 	import { T } from '@sveltevietnam/i18n/runtime';
 	import type { Message } from '@sveltevietnam/i18n/runtime';
-	import type { Status } from '@sveltevietnam/kit/constants';
 	import { Contexts } from '@sveltevietnam/kit/contexts';
-	import { REDO_COMMAND, UNDO_COMMAND } from 'lexical';
 	import { onMount } from 'svelte';
-	import { type ClassValue } from 'svelte/elements';
 	import { on } from 'svelte/events';
 
-	import * as m from '$data/locales/generated/messages';
-
 	import { Editor, type EditorInit } from './editor';
-	import ToolbarBlockTools from './editor/components/ToolbarBlockTools.svelte';
-	import ToolbarInlineTools from './editor/components/ToolbarInlineTools.svelte';
-	import { CalloutStatusDropdown, FORMAT_CALLOUT_COMMAND } from './editor/plugins/callout';
+	import Toolbar from './editor/components/Toolbar.svelte';
 
 	export interface RichTextEditorProps extends EditorInit {
 		placeholder?: Message<'string', never>;
@@ -33,89 +26,10 @@
 			onchange?.((event as CustomEvent<{ html: string }>).detail.html);
 		});
 	});
-
-	function undo() {
-		editor.lexical.dispatchCommand(UNDO_COMMAND, undefined);
-	}
-
-	function redo() {
-		editor.lexical.dispatchCommand(REDO_COMMAND, undefined);
-	}
-
-	function help() {}
-
-	function handleSelectStatusDropdown(status: Status) {
-		editor.lexical.dispatchCommand(FORMAT_CALLOUT_COMMAND, { status });
-	}
 </script>
 
 <div class="border-onehalf bg-surface relative flex flex-col-reverse border-current">
-	<!-- toolbar -->
-	<div
-		class="border-y-onehalf bg-surface-subtle z-2 sticky bottom-0 flex items-center gap-0.5 overflow-visible border-current p-0.5"
-		role="toolbar"
-		aria-labelledby="rich-text-toolbar-label"
-	>
-		<p class="sr-only" id="rich-text-toolbar-label">
-			<T message={m['components.rich_text_editor.toolbar.name']} />
-		</p>
-
-		<!-- add tooltip on hover with more info for each action -->
-		{#snippet toolbarAction(
-			iconClass: string,
-			sr: Message<'string', never>,
-			onclick: () => void,
-			active?: boolean,
-		)}
-			<button class={['c-btn c-btn--icon p-2', active && 'bg-primary/50']} {onclick} type="button">
-				<i class="i {iconClass} h-5 w-5"></i>
-				<span class="sr-only">
-					<T message={sr} />
-				</span>
-			</button>
-		{/snippet}
-		{#snippet separator(classes?: ClassValue)}
-			<div class={['bg-outline flex w-px shrink-0 self-stretch', classes]}></div>
-		{/snippet}
-
-		<!-- undo -->
-		{@render toolbarAction(
-			'i-[ph--arrow-counter-clockwise]',
-			m['components.rich_text_editor.toolbar.undo'],
-			undo,
-		)}
-		<!-- redo -->
-		{@render toolbarAction(
-			'i-[ph--arrow-clockwise]',
-			m['components.rich_text_editor.toolbar.redo'],
-			redo,
-		)}
-
-		{@render separator('ml-2')}
-
-		<ToolbarBlockTools {editor} />
-
-		{@render separator('mr-2')}
-
-		<ToolbarInlineTools {editor} />
-
-		{@render separator('ml-auto')}
-
-		{#if editor.block.type === 'callout'}
-			<CalloutStatusDropdown
-				status={editor.block.props.status}
-				onselect={handleSelectStatusDropdown}
-			/>
-			{@render separator()}
-		{/if}
-
-		<!-- open help dialog -->
-		{@render toolbarAction(
-			'i-[ph--question-mark]',
-			m['components.rich_text_editor.toolbar.help'],
-			help,
-		)}
-	</div>
+	<Toolbar {editor} />
 
 	<div class="z-1 relative">
 		<!-- composer -->
